ScheduleConfigurationScreen -> MVVM
